requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itchat==1.3.10
lxml==4.3.5
numpy==1.17.0
pandas==0.25.0
parsel==1.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.0
pytz==2019.1
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.2
selenium==3.141.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
TA-Lib==0.4.17
tushare==1.2.39
Twisted==19.2.1
urllib3==1.25.3
w3lib==1.20.0
zope.interface==4.6.0
